docs(server): replace `mkdocs` with `pdoc` & new CI action
