+++ .gitignore
@@ -299,3 +299,5 @@ crashlytics-build.properties
 .ionide
 
 # End of https://www.toptal.com/developers/gitignore/api/macos,visualstudiocode,python,unity
+
+website/docs/*
