+++ python/arflow/__init__.py
@@ -1,3 +1,11 @@
+from arflow.core import *  # noqa
 from arflow.replay import *  # noqa
 from arflow.serve import *  # noqa
 from arflow.service_pb2 import *  # noqa
+
+# https://pdoc.dev/docs/pdoc.html#exclude-submodules-from-being-documented
+__all__ = [  # noqa
+    "core",  # noqa
+    "replay",  # noqa
+    "serve",  # noqa
+]
