--- python/arflow/scripts/start_service.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-PYTHONPATH=./service/schema:$PYTHONPATH python service/main.py
