--- python/docs/index.md
@@ -1,1 +1,0 @@
-# ARFlow
