--- python/mkdocs.yml
@@ -1,5 +1,0 @@
-site_name: ARFlow
-site_url: https://cake.wpi.edu/ARFlow/
-nav:
-  - Home: index.md
-theme: readthedocs
