--- python/scripts/launch_example_simple_replay.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-poetry run python examples/simple_replay.py
