--- python/scripts/launch_example_simple_server.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-poetry run python examples/simple_server.py
