--- python/scripts/make_docs.sh
@@ -1,11 +1,0 @@
-#!/bin/sh
-
-mkdocs build
-
-if [ $? -ne 0 ]; then
-    echo "mkdocs build failed"
-    exit 1
-fi
-
-mkdir -p ../website/docs
-cp -r site/* ../website/docs
